+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Revert Loki to major version 2 because upgrading to version 3 required manual intervention as Loki
   refuses to start. So until this issue is tackled, reverting is the best immediate fix.
   See https://github.com/NixOS/nixpkgs/commit/8f95320f39d7e4e4a29ee70b8718974295a619f4
+- Add prometheus deluge exporter support. It just needs the `shb.deluge.prometheusScraperPasswordFile` option to be set.
 
 ## Other Changes